preserve context in logging of predictive scaling
We keep some Procs around for doing predictive scaling analysis that are
passed on to the scaledown routines, that are called to prevent scaling
down when doing that would then cause us to immediately scale back up
because of predictive scaling.

Those Procs did not have context to log where they came from when called
outside of predictive scaling, so this just captures the context in a
string and logs it.

Change-Id: If6a8a20e3fdd63060ce434cce5f7342f8d7d2ef3